requirements.txt
azure-functions
azure-keyvault-secrets
azure-identity
httpx
